+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/components/CaseCreationForm/hooks/useCaseCreationForm/useCaseCreationForm.ts
@@ -1,11 +1,11 @@
-import { TWorkflowDefinition } from '@/domains/workflows/fetchers';
+import { TWorkflowDefinitionById } from '@/domains/workflow-definitions/fetchers';
 import { useCreateWorkflowMutation } from '@/domains/workflows/hooks/mutations/useCreateWorkflowMutation/useCreateWorkflowMutation';
 import { createContextFromFormData } from '@/pages/Entities/components/CaseCreation/components/CaseCreationForm/hooks/useCaseCreationForm/utils/create-context-from-form-data';
 import { useCaseCreationContext } from '@/pages/Entities/components/CaseCreation/context/case-creation-context/hooks/useCaseCreationContext';
 import { AnyObject } from '@ballerine/ui';
 import { useCallback } from 'react';
 
-export const useCaseCreationForm = (workflowDefinition: TWorkflowDefinition) => {
+export const useCaseCreationForm = (workflowDefinition: TWorkflowDefinitionById) => {
   const { isLoading, mutateAsync } = useCreateWorkflowMutation();
   const { isMultipleCasesCreation, setIsOpen } = useCaseCreationContext();