+++ packages/common/src/utils/index.ts
@@ -17,4 +17,5 @@ export { raise } from './raise';
 export { log } from './log';
 export { replaceNullsWithUndefined } from './replace-null-with-undefined';
 export { dump } from './dump';
+export { isNonEmptyArray } from './is-non-empty-array';
 export { type IErrorWithMessage } from './is-error-with-message';