+++ services/workflows-service/src/workflow/workflow.controller.internal.unit.test.ts
@@ -38,6 +38,12 @@ class FakeEntityRepo extends BaseFakeRepository {
   }
 }
 
+class FakeUiDefinitionService extends BaseFakeRepository {
+  constructor() {
+    super(Object);
+  }
+}
+
 export function buildWorkflowDefinition(sequenceNum: number, projectId?: string) {
   return {
     id: sequenceNum.toString(),
@@ -80,6 +86,7 @@ describe('WorkflowControllerInternal', () => {
   let userService;
   let salesforceService;
   let workflowTokenService;
+  let uiDefinitionService;
   const numbUserInfo = Symbol();
   let testingModule: TestingModule;
 
@@ -100,6 +107,7 @@ describe('WorkflowControllerInternal', () => {
     userService = new FakeEntityRepo();
     salesforceService = new FakeEntityRepo();
     workflowTokenService = new FakeEntityRepo();
+    uiDefinitionService = new FakeUiDefinitionService();
 
     eventEmitterSpy = {
       emitted: [],
@@ -123,6 +131,7 @@ describe('WorkflowControllerInternal', () => {
       userService,
       salesforceService,
       workflowTokenService,
+      uiDefinitionService,
     );
     const filterService = {} as any;
     const rolesBuilder = {} as any;