+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ballerine/workflow-node-sdk
 
+## 0.5.39
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.39
+
+## 0.5.38
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.38
+
 ## 0.5.37
 
 ### Patch Changes